requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
pandas==2.1.4
openpyxl==3.1.2
requests==2.31.0
python-dotenv==1.0.0
openai==1.3.0
aiohttp==3.9.1
